users can now log in and out.
search bar added but needs to be improved.
user needs to be able to search for other users.
user needs to be able to add friends.
user needs to be able to see friends list.
user needs to be able to see friend requests.
user needs to be able to send friend requests.
user needs to be able to accept friend requests.
user needs to be able to reject friend requests.